% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc.R
\name{LMA.Function}
\alias{LMA.Function}
\title{基于GCV的最小二乘模型平均}
\usage{
LMA.Function(y, xmatrix, model.list = NULL, use.func = my_RFunction)
}
\arguments{
\item{y}{n*1的向量，自变量}

\item{xmatrix}{n*p的矩阵，因变量}

\item{model.list}{列表，即list(v1,v2,...,vM)，列表元素为向量，每一个向量代表一个模型，vi对应的模型使用的自变量为 xmatrix[,vi]；默认为 NULL，等价于 list(c(1),c(1,2),...,c(1,2,...,p))}

\item{use.func}{函数，使用的最优化函数，my_RFunction 或 my_CppFunction}
}
\value{
M*1的向量，即model.list中模型对应的权重
}
\description{
最小化损失函数求模型平均的权重
}
