% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{my_CppFunction}
\alias{my_CppFunction}
\title{使用Rcpp的外罚函数法}
\usage{
my_CppFunction(
  x0,
  lossf,
  func1,
  func2,
  epsilon = 1e-06,
  sigma = 0.5,
  gamma = 1.1,
  MAX_ITER = 1000L
)
}
\arguments{
\item{x0}{初始值}

\item{lossf}{目标函数}

\item{func1}{步骤一函数}

\item{func2}{步骤二函数}

\item{epsilon}{算法参数}

\item{sigma}{算法参数}

\item{gamma}{算法参数}

\item{MAX_ITER}{最大迭代次数}
}
\value{
最优值
}
\description{
使用Rcpp的外罚函数法
}
