% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfunc.R
\name{my_RFunction}
\alias{my_RFunction}
\title{使用R的外罚函数法}
\usage{
my_RFunction(
  x0,
  lossf,
  need.func1 = func1,
  need.func2 = func2,
  epsilon = 1e-06,
  sigma = 0.5,
  gamma = 1.1,
  MAX_ITER = 1000
)
}
\arguments{
\item{x0}{初始值}

\item{lossf}{目标函数}

\item{need.func1}{步骤一函数}

\item{need.func2}{步骤二函数}

\item{epsilon}{算法参数}

\item{sigma}{算法参数}

\item{gamma}{算法参数}

\item{MAX_ITER}{最大迭代次数}
}
\value{
最优值
}
\description{
使用R的外罚函数法
}
